Moving ColumnType component
