+++ amundsen_application/static/js/pages/TableDetailPage/ColumnList/ColumnType/index.tsx
